# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Drive API
  description: |
    A secure file sharing platform that allows users to upload, expose, and manage access to their files.
    
    ## Features
    - User authentication with phone numbers
    - File upload and management
    - Secure file exposure with time limits
    - Access request and approval system
    - IP-based access control and blacklisting
    - Premium phone-based sharing with secure tokens
    
    ## User Types
    - **Free Users**: Can expose one file at a time for maximum 5 minutes
    - **Premium Users**: Can expose multiple files with custom expiration times up to 24 hours, phone-based sharing
    
    ## Authentication
    Most endpoints require Bearer token authentication. Obtain tokens through the login endpoint.
    
  version: 1.0.0
  contact:
    name: Drive API Support
    url: https://expose.tech/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.expose.drive.tech
    description: Production server
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register new user
      description: Register a new user with a phone number. Returns a login hash that must be used for authentication.
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: Phone number in international format
                  example: "+918043652258"
            examples:
              valid_phone:
                summary: Valid phone number
                value:
                  phone: "+918043652258"
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user_id:
                    type: integer
                    example: 1
                  login_hash:
                    type: string
                    description: Unique hash required for login
                    example: "abc123def456..."
                  phone:
                    type: string
                    example: "+918043652258"
        '400':
          description: Invalid phone number format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User with this phone number already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: User login
      description: Authenticate user with phone number and login hash obtained during registration.
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, login_hash]
              properties:
                phone:
                  type: string
                  description: Phone number used during registration
                  example: "+918043652258"
                login_hash:
                  type: string
                  description: Login hash received during registration
                  example: "abc123def456..."
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT token for API authentication
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type:
                    type: string
                    example: "bearer"
                  user_id:
                    type: integer
                    example: 1
                  user_type:
                    type: string
                    enum: [free, premium]
                    example: "free"
                  login_hash:
                    type: string
                    example: "abc123def456..."
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/upgrade-premium:
    post:
      summary: Upgrade to premium
      description: Upgrade current user to premium (for testing purposes)
      tags: [Authentication]
      responses:
        '200':
          description: Successfully upgraded to premium
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully upgraded to premium"
                  user_type:
                    type: string
                    example: "premium"
                  phone:
                    type: string
                    example: "+918043652258"
        '400':
          description: User is already premium
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # File Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the login endpoint
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "File not found"
        code:
          type: string
          description: Error code
      required:
        - detail
...